Make cache restore conditional on previous cache being restored
The build of NAS2D depends on `vcpkg` packages. If `vcpkg` packages have changed, then we likely need to rebuild (almost all of) NAS2D.

Similarly OPHD depends on NAS2D (and on `vcpkg` packages). If NAS2D couldn't be restored, we likely need to rebuild (almost all of) OPHD.