Use env vars for cache keys when saving
Previously we used the key from the restore step as a way of ensuring the save key would match, even if the restore key was modified. Now we have an env var with the value, so we can use that to ensure consistency between restore and save.

In some cases we may skip the restore step, in which case the restore step won't have a cache key set. In those cases it's better to use the env var, which will always be set.