Remove component member `disconnect` calls from `GameState` destructor
These are being called on component objects, which contain the corresponding `Signal` objects. When the `GameState` is destructed, it will destruct all component objects, which will in turn destruct their component `Signal` objects. As part of that, all listeners will be disconnected. Thus there was little point in making manual calls to `disconnect`.