Make `ListBoxBase::clear()` `protected`
We are going to split this function, so the base class implementation is not fully functional. This we don't want client code maing direct calls to this method. We only want to allow derived classes to use it. Derived classes can of course provide their own `public` method with a complete implementation.