Remove `mMainReportsState` member `disconnect` calls
This case is a little less clear due to extra levels of indirection, and the use of `static` objects. When `GameState` is destructed, the `MainReportsUiState` member field is destructed. In the destructor of `MainReportsUiState`, it destructs each of the component `UiPanel` objects pointed to by the `static` `Panels` array. The `takeMeThereSignal` comes from those `UiPanel` objects. The `UiPanel` objects derive from `ReportInterface`. That interface is what holds the `TakeMeThere` `Signal` objects. They are destructed when the `ReportInterface` class is destructed, which are what the `UiPanel` fields point to. Thus the `Signal` objects contained by `ReportInterface` will have all listeners disconnected.