Update Ming build environment
It seems the `-Wno-dangling-reference` flag is needed to compile with this update. It seems the new flag is buggy and produces false positives.

The env vars for `-Wno-redundant-decls` and `LDFLAGS_EXTRA` are not strictly necessary with this build image, since they are specified in the build image. Though for someone interested in building this project themselves, they might not think to look there. Placing them here makes for easier to follow build instructions.