Dynamically specify build env vars for MacOS
Depending on the version of the runner, we can get different versions of `brew`, which may install files to different paths. To avoid problems with changing paths, we can query `brew` for the needed paths.

Note that `CPLUS_INCLUDE_PATH` is a compiler specified env var, while `LDFLAGS_EXTRA` is a custom var used by the `makefile`.