Add `protected` method `buildSpecificStringTable`
It's a bit strange having to pass in the `position`, though it does mean the result can be stored in a `const` variable. It also means we don't need to store extra data in the class instance to handle this.

Long term we should probably make both tables fields of the class, rather than rebuild them every frame. Though doing so would mean moving the `#include` for `StringTable.h` to the `StructureInspector` header file. That has transitive include implications, which I'd like to avoid for the time being.