Update margins to standard sizes
Note the second constant `20` is additional margin between sections. It is relative spacing between the end of the generic table and the start of the specific table. It's not related to the title bar height. Though possibly, the large gap between sections was originally set that large because of a mistaken assumption the title bar height needed to be included in the calculation.

Original section offset set in commit: `1ba57eac40858c4998124e783a4c43839363baa4`