Remove unused `OnMoveSignal` from `Control`
This feature was unused, and seemed unlikely to ever be used. A `Control` is generally told where it is positioned or to be moved by it's owner. As the owner is already in control of this, there is no need for a `Signal` to be informed of it. Better to reduce bloat from this common base class.