Remove `Control::rect() const`
Replace uses with `Control::area() const` uses.

It was tempting to go the other way, since `rect()` was more common than `area()`. The opposite was true for the `area` setter method, which had no corresponding `rect` setter method. Overall, I think `area()` reads better than `rect()` where it's used. Of note, `area()` corresponds well with `position()` and `size()`, which we are already using. It would have been weird to name those other methods `point()` and `vector()`, so it seems odd to use the name `rect()`. The name should be a description of use, rather than a description of the data type used.