Convert console message to exception in `ProductCatalogue.cpp`
In general, we should prefer using exceptions to report errors rather than rely on console messages.

In this particular case, we are reporting an issue with a static data asset shipped with the game. If that's wrong, we should fail loudly. This type of error should be caught and fixed before shipping a release.