Use `using` to re-export `update` with `public` access
The original access was `protected`. That worked within NAS2D, where there is a `friend` declaration to give access to `State::update` to `StateManager`. Outside of NAS2D, where NAS2D has no knowledge, and can not declare a `friend`, some other means was needed to obtain access. This was done previously by creating an alias for the function with `public` access. However, it is possible to give the original method a new access level in the derived class, without having to create an alias.