Rename `mClose` to `mCancel`
An informational window might reasonable have a "Close" button for when the user is done viewing. For an operation window a "Cancel" button makes a bit more sense.

For mixed usage windows, with information and ability to perform operations, then perhaps treat it as an informational window. Any potential operations may depend on the information presented.