Make `StructureListBox::setSelected` clear selection on invalid value
When the list is empty, any value is invalid. Additionally there is really no speed benefit to an early check and exit in that case.

Potentially we could also eliminate the special check for `nullptr`, if we assume no value in the table will be equal to `nullptr`. Speed wise there may still be a benefit to an early return on `nullptr` if the table is large.