Simplify calling `lstMineFacilities.setSelected`
It is safe to pass `nullptr`, and doing so will set the `ListBox` to `NoSelection`.

Calling `setSelected` will trigger the `selectionChanged` event. This in turn will call `MineReport::onMineFacilitySelectionChange` which will call `updateManagementButtonsVisibility`. Hence there is no need to manually call the last method. Doing so will simply make it run twice. The method is idempotent, so the second run will have no observable effect.