Position `mPopulationPanel` so it stays on screen
For larger font sizes, the window can enlarge. Potentially it's position can allow it to extend beyond the space available to the main window, causing the contents to be clipped. This can happen when the main window is set near the minimum size, and the font sizes are increased by 4 points. This change detects such an overflow and repositions the population panel so it remains fully on screen.