Split `Music` to separate cache file
The `<memory>` header can be a bit of an expensive include in terms of compile time. Couple that with how frequently `Cache.h` is included, and how few of those includes need the entry that makes use of `<memory>`, and it becomes a good candidate to split to a separate file. Additionally there is a NAS2D issue that may require substantially increasing the transitive includes from `Music.h` to include `<SDL.h>`, and there's even more incentive to split that off.