Update CircleCI build config
Disable one GCC/Mingw warning flag that was enabled as part of a group and seems to be buggy.

Disable two GCC/Mingw warning flags which we are not yet ready to enable. These were not specified previously for OPHD, though are now being used in NAS2D.

Disable several Clang warning flags which we are not yet ready to enable. These were not specified previously for OPHD, though are now being used in NAS2D.