Remove CI flag `-Wno-dangling-reference`
This is now disabled by `#pragma` in the code.

Using `#pragma` makes it easier to deal with switching compilers locally. The new warning is enabled by `-Wextra`, which is supported by both GCC and Clang. However, the flag to disable that specific warning, `-Wno-dangilng-reference` is only supported by GCC. This makes it awkward to set flags to enable as many warnings as possible, and not cause errors when switching between compilers.