Update `operational` check to use `commRange > 0`
When a structure is not operational, `commRange()` will return `0`, so the new code includes the old condition. Additionally, we avoid treating every structure as having communication range for the tile it's on. Imagine a structure built far away under a communications tower, and then having the tower be destroyed. Technically the tile for that structure is at distance 0 from the structure, so could have been considered under communications range. That would allow robots to operate on that one tile. This closes that loophole.