Add `isInGridArea` check to `translateCoordsToIndex`
This allows for simplified calling. There is no longer a need to ensure the `position` is within the bounds of the control, and within the grid area of the control. This eliminates concerns about negative offsets and overflow. It also allows for quick return of `NoSelection` if the position is outside of the grid area.