Remove no longer needed check for `isInGridArea` in `OnMouseDown`
Technically there is a slight behavior change here, though it may be considered fixing a slight bug. Normally when clicking in the control and not on an icon, it would result in a `NoSelection` choice. However, if the control bounds extended beyond the grid area, the click would simply be ignored, creating inconsistency in behavior. With this change, a click anywhere withinthe control bounds, which is not on an icon will result in a `NoSelection` choice.

Note that this change may result in extra `NoSelection` events being raised to the enlarged area of clicks for making a `NoSelection` choice. In particular when clicking beyong the last row or column that could have been filled (whether or not it contains any icons).