Remove calls to `toLowercase` in `IconGrid`
Using `toLowercase` can make sense for user facing interfaces. It makes less sense for internal functions, particularly ones using hardcoded constants in a consistent manner. It may even be undesireable to use `toLowercase` in cases where someone might want to be able to discriminate based on case. Not to mention it adds to the runtime overhead to do a case insensitive compare.