Switch check to guard against `NoSelection`
If the list is empty, then there must be `NoSelection`. Hence the updated check guards against the old condition of the list being empty. Additionally, if there is `NoSelection`, we can't really increment or decrement sensibly. In particular, the decrement operation was unsafe when there was `NoSelection`, producing a garbage selection result that didn't match `NoSelection`. The updated code avoids the unsafe case.