Move unselect guard first in `MapViewState::onStructuresSelectionChange`
We only need to clear the other `IconGrid` instances when a new object is selected. The clearing is to ensure at most one object is selected. If there is no structure selection then there is no need to clear the other selections.

We also want to avoid mutual clearing should `IconGrid` be updated to raise the changed selection event when selections are cleared. Calling clear selection on other `IconGrid` objects could cause another changed selection event to occur which then clears the selection that raised the initial event.