Avoid clearing other selections when no robot selection
Assuming `IconGrid` may eventually raise selection changed events when selections are cleared, we want to avoid triggering mutual clearing, and infinite recursion. Unfortunately, in this case, the `clearMode()` call does trigger clearing of other `IconGrid` selections.

Avoid setting later in the function that would be inappropriate if the robot selection was being cleared.