Uncouple `clearSelections()` from `clearMode()`
The majority of cases where `clearMode()` is called, a call to `clearSelections()` would not be necessary. In most cases there is modification of the `IconGrid`, which would automatically clear any selection. In the remaining two cases, there is a call to `clearSelections()` immediately after the call to `clearMode()`.

By uncoupling the two, it makes it easier to avoid unwanted interactions between `IconGrid` instances where a change to the selection of one `IconGrid` causes a change to the selection of another `IconGrid`. This might be selecting in one clears others, which causes the initial one to be cleared as well, or infinite recursion of changes to unselect the other's selection.