Remove unused `DieSignal` from `MapObject`
If we really do need it in the future, we can add it back as a `Delegate`, rather than a `Signal`. Though if we only need to notify game objects, we could also use a direct call to them. If we develop a root `Game` object, as part of `libOPHD`, then `MapObject` would be allowed to know about it, and could make direct calls to it. Perhaps we could pass in a reference to the `think` function.