Add `MainReportsUiState::onTakeMeThere` intermediate
All reports now send the event to the `MainReportsUiState::onTakeMeThere` object, which can then distribute the `Signal` to any listening object. This hides the detail of multiple internal reports from users of `MainReportsUiState`.

Normally it would be more efficient to pass handlers through, without an intermediate. Though there's a construction order problem here, due to the `static` `Panel` instances, which are constructed some time after `MainReportsUiState` is constructed. This prevents feeding `Delegate` objects from the `MainReportsUiState` constructor through to the individual reports.

Additionally, the intermediary may provide a place for us to silence events from reports that are not currently visible, but may still be responding to input events as if they were. This could help with an open bug.