Move `die` method from `MapObject` to `Robot`
The `die` method is only ever used with `Robot`. The `Structure` class had an override that throws an exception so it will fail loudly if it was ever used on a `Structure`. Even better than a runtime error, is a compile time error. By moving the function out of `MapObject`, and not providing such a method in `Structure`, any attempt to reference the name on a `Structure`, or rather on a non-`Robot` will fail at compile time.