Move `processTurn()` from `MapObject` to subclasses
Although all `MapObject` classes may have turn processing, they are never handled together as one collection, so there is no need for a `virtual` method on the base class. Additionally, we want to vary the parameters passed to the different subclasses when processing them, so it no longer makes sense to use a common method.