Make `onTaskComplete` pure virtual in `Robot`
There were no pure virtual methods in `Robot` previously, so this change makes it an abstract class, which it probably already should have been.

The subclasses now have a method that is capable of modifying the `TileMap` when a task is complete. This will allow code to be moved from `MapViewStateEvent.cpp` into the `Robot` subclasses. At least it will, once they know where they are on the map.