Remove comments "only here for loading games"
That may be true, though the code is changing and there may be changes coming soon when that is no longer true. Plus the comments are now of questionable relevance. There was some oddities of initialization previously, which were due to how loading games worked. Now though, things are more consistent, so the difference doesn't really matter.