Prefer `{}` over `""` for empty `std::string` initialization
It's easier for the compiler to default initialize a `std::string` to empty, rather than copy initialize it from a C-string, which may require scanning it to find the length, and copying every character. With a C-string copy, the compiler may not be able to see the string is empty and optimize accordingly.