Fix Segmentation fault bug
It seems the `CheckBox` click handler was being activated even though the `MineOperationsWindow` was not visible, which was running the `MineOperationsWindow` click handling method when `mFacility` was `nullptr`.

With the click handling collapsed to a single function, it's easy to bug fix in one place, rather than insert checks into 4 separate methods.