Use `TileInspector::onVisibilityChange` to hide when no tile to show
This is better than waiting until the `update` or `draw` code to detect the problem and correct the visibility. For one, setting `visible` is a self modifying operation, so can't be placed in a `const` function, such as `draw`. For another, if we wanted to detected this as an error, we should do so at a place when the call that set the visibility is still on the call stack.