Fix multi-line `ToolTip` Y-offset
First try to place the `ToolTip` above the `Control`, and if that doesn't fit, place it below the `Control`.

The old code to place the `ToolTip` below the `Control` incremented by the `ToolTip` size rather than the `Control` size. This wasn't so noticeable when the two were both about one line of text in height, though is more noticeable with multi-line text.