Remove debug code to print last modification times
It actually causes an error when there is no cache to restore, since it tries to list directories that don't exist in that case. Plus, this code was meant to be temporary. Additionally, it did not reveal the cause of the rebuilds, since it showed cached build outputs as newer than the source files, yet rebuilds still happened.