Remove use of `std::optional`
I kind of like the idea of using something like `std::optional` for single function operation. Though given the code structure, having two result functions (`crashed` and `crashedLanders`), it seems a bit redundant here. There is already a check for `crashed()`, which makes the `has_value()` check redundant. Additionally a `std::optional` field will take at least as much space as the value it contains, and probably more to distinguish when it is empty. Given the current design, and the style used elsewhere in the code, I think it makes the most sense to keep the two result functions (the validity checking function, and the data returning function), and eliminate the use of `std::optional`.