Check validity of `ProductType` before adding name to `ProductListBox`
Lookup of an invalid `ProductType` would result in an exception, which would crash the program. Not fun. Instead, display text saying the `ProductType` is invalid. This should hopefully still be an obvious error, without destroying the game experience.

This also provides a way to test certain `ProductListBox` and `ListBoxBase` features. Instead of checking if `productCount > 0`, we can check for `productCount >= 0` to add more entries to the `ProductListBox`.