Rename variable `firstInvisibleIndex` for clarity
With the old name, the use of `std::min` with `mItems.size()`, and the `<` loop condition, it looked like there may have been an off by one error. However the result does appear to be a pixel perfect visibility calculation. The renamed field is actually the index of the item that would be just out of view.