Add `ComboBox` to `demoLibControls`
Note: At present there is an uncaught exception which can be triggered when clicking on a `ComboBox` list to select an item, and the `width` of the `ComboBox` is too narrow to display the full string. The problem occurs in `TextField::onMouseDown` when `substr` is called with an invalid `scrollOffset` value.