Add `smallChange` setting to `ScrollBar`
The `smallChange` value is the delta when clicking an increase or decrease button, or from one click of the mouse scroll wheel. This is useful when the `ScrollBar` is used in a context where the `value` is the pixel offset to scroll an area, and it's desired to scroll by an entire item size in pixels. The track thumb can still be used for pixel level scrolling. If not specified, the `smallChange` setting defaults to `1`, which matches the old behavior.

This feature is used by `ListBox` to move by one `itemSize.y` at a time, rather than 1 pixel.