Split early return checks
Add check if click is within control to primary return check. We don't handle clicks when they are not within the control.

Eventually we will allow some click handling outside of the scroll area, but still within the control, namely for right clicks. Hence we want to preserve the original scroll area check, though it will be moved down to later in the method.