Style update for `ListBoxBase` drawn items
This makes the item borders appear similar to previous code. With the previous over drawing, the `inset(2)` left a gap of 1 pixel between the item border and the `ListBoxBase` border. It also left padding of 2 + 2 pixels between items. With this update, we shrink the padding from 2 to 1 pixels, leaving a 1 pixel gap between the item border and the `ListBoxBase` border (restored behavior), while leaving a gap of 1 + 1 pixels between items (different from previous). This produces a more balanced look, with more consistent spacing in the different directions.