Clearn highlight selection when mouse not over scroll area
Previously the `ListBoxBase` border region was considered for tracking of highlighted items. This was particularly odd when a scroll bar was present, as that area was specifically excluded, and was between the item and the border. It was odd to see the item highlighted when the mouse was on the far side of the scroll bar, over the border, but not when over the scroll bar. With this change, the mouse has to be over an item itself, not simply somewhere over the list box control that isn't the scroll bar.