Auto set `ListBox` focus on mouse over
This allows the scroll wheel to function immediately upon mouse over.

As an aside, this also fixes the issue of the `ListBox` not losing the focus highlight when clicking off onto a `RadioButtonGroup`. As soon as the mouse leaves the `ListBox` it loses focus, so the problem is solved before any other `Control` is clicked.