Fix highlight to index tracking in `ListBoxBase`
There was a subtle off by one bug here. It was using a relative offset to the `ListBoxBase` position, rather than the scroll area of the `ListBoxBase`. That meant the `ListBoxBase` border was the first pixel for highlighting, which was excluded in an earlier check, however it meant tracking for the end of each item and the start of the next was also off by 1 pixel. Using the `mScrollArea`, inside the border, rather than the `Control` position fixes this offset problem.