Add `hasFocus()` check to `onMouseDown`
The `hasFocus` property is set earlier, based on whether the click is within the control's bounds. We should refuse to process mouse clicks that are not within the control's bounds.

This fix seems to prevent most crashes, though it is not the only source of problems. There is some inappropriate uses of a pixel offset value as if it was a character index offset value, which seems to be the real source of problems. Additionally, it was found before that clicking slightly before the end of the string visually would cause crashes. This seems to relate to `offsetX` being relative to the `Control` position, rather than the padded offset to where `text` is drawn. The check to clip handling to the bounds of the control reduces the area where such a click can happen.