Fix recently introduced bug when clicking last character
The bug was causing the last character to be ignored, so the cursor would end up being placed before the second last character, rather than before the last character.

It should be noted the `index` value is used as a sub-string length, rather than an index. Notably, the selected index at the end is `index - 1`, as it's a length value rather than an index.