Adjust loop starting condition
A sub-string of length 0 will have pixel width of 0. A pixel width of 0 will never exceed the `virtualOffsetX`, which is at least 0.

Additionally, we set an index value of `subStringLength - 1`, which we never want to be negative. Hence, by starting the loop at `1` instead of `0`, we ensure the computed index value can never be negative.