Check for `isMineFacility()` rather than `active()` in `MiniMap`
When drawing the `OreDeposit` locations on the `MiniMap`, and choosing the color/symbol to draw, check for the absence of a `MineFacility`, rather than the absence of the `active` flag. The `active` flag is only ever set when a `MineFacility` is constructed.

Technically, this may be more accurate than previous, since the `active` flag is never cleared. If a `MineFacility` were destroyed, the highlighting should revert back to the inactive state.