Verify the `StructureType` collection order
Existing code required the `StructureType` collection to contain an entry for every built-in `StrutureID` enum value. It allowed the entries to be in any order, and provided a lookup `map` to convert between them.

The updated code will soon `static_cast` between `StructureID` and the `StructureType` index values. For that to happen, the collection must be in the same order as the `StructureID` enum values. Hence we add this stricter verification here.

Currently the `StructureType` data has been updated to match 1-1 with the `StructureID` enum values, so this check is expected to pass as written.

Of course, testing this check required temporarily modifying the `StructureType` data, to ensure the error was detected.