Move item border drawing from `ListBoxBase` to subclasses
This is really a decision that should be made in the subclasses. It just so happens that all the current subclasses use the same border style. Though that was sort of imposed by putting the item border drawing in the base class.

This makes the `ListBoxBase` draw code more consistent with `ListBox`. It also potentially allows `ListBox` draw code to inherit from `ListBoxBase`, though support for list box themes would need to be moved to `ListBoxBase` first.