Set combo box title to align bottom with first button row
Currently the font text height and the filter button sizes are the same. That means at current sizes the output will not move from the current position. However, as the filter button size is increased, the title for the `ComboBox` will move down to remain bottom aligned with the first row of filter buttons.