Shorten toggle button tool tip text
It's a toggle button. It behaves like a toggle button. The user will very quickly figure out it's a toggle button without being told that part. As long as the description look non-threatening, they will click it and find out what it does. For example, text stating something is an "Overlay" would imply it's an informational button.