Add `Button::click()` method
Useful to simulate clicking the `Button`. Potentially control containers can use this to trigger event handlers, without having to store a separate copy of the handler. There is currently no accessor for the set handler, so no way to retrieve it and call it directly. Thus the only way to call the handler, without having made a copy, is to ask the `Button` to call it.

Potentially this may also have uses for user interface test code.