Allow all `clickHandler` parameters to be optional
When someone is first designing an interface, they might not have click handler functions setup yet, and just want to see how things look. In such as case, it makes sense to allow for the creation of non-functional buttons. They could always have passed an empty `Delegate` for the parameter, so creating non-functional buttons was always a possibility.

Additionally, by setting the last parameter as optional, we can make earlier parameters optional too.