Add `ListBoxBase::setSelectedInternal` helper method
There is no bounds checking on this internal method. It can be used with either a valid `index` value or with the special `NoSelection` sentinel value.

In all cases, if the set value is different from current, the current selection is updated, and any needed selection changed event is raised.