Group clearing highlight with clearing selection
We probably want clearing the selection to be last. Clearing the selection may involve calling the selection changed event, which passes control to external code. We want the `ListBoxBase` instance to be in a consistent state when that happens, with all class invariants holding. That includes any index value being either valid or `NoSelection`. We likely also want the `ListBoxBase` scroll offset to be within a valid range for it's size and contents.