Extract method `onDropdownButtonClick()`
This should really be called by the `Button` delegate, rather than manual hit testing in `ComboBox::onMouseDown`. Though that would be a behavior change, since the entire `mBarRect` is valid for opening and closing the list. The `TextField` is display only. The entire thing works like the drop down `Button`.