Add `mMinHeight` field to `ComboBox`
When `TextField` and `Button` are constructed, they are set to an appropriate minimum size. We record the larger of the two `height` properties and use this to enforce a minimum size on the `ComboBox`. Note that the `TextField` and `Button` can increase in `height` if the `ComboBox` is resized to be larger, hence why we record the initial value.