Remove unused set method `ComboBox::text`
This raises too many questions, and is probably not something we want to support. Better to just remove this feature than try to fix it.

For one, the method could deliver the changed event twice. Once from the manual call, and another possible one through the `ListBox::selectIf` method, assuming it finds a match. What happens if it doesn't find a match is questionable. Only one change event is raised, the `ListBox` keeps it's previous selection, and the text field changes, meaning it no longer matches the `ListBox` selection.

We kind of just want a basic `DropDownList`, not really a full `ComboBox` with editing capabilities.