Increment cursor index by the length of inserted text
We would expect the length to typically be 1, though that might not be the case. The SDL documentation suggests this could be up to 32 characters in one chunk. That may be the case with Input Method Editors (IMEs), which might assemble multiple characters before returning them all as a chunk.