Remove the CircleCI package step too
If we're not using the artifacts, we also don't really need to package them. This means we're testing less of the build process with each run. Granted the packaging part almost never gets any attention or changes, so we probably don't need to test it by running it on every run. If we make changes, we'll know the test the changes when we make them.