Remove `depth` guard
This was necessary before when the `excavated` call wasn't under the `productionTime` guard, since `Robot` placement on the map, and having valid `MapCoordinate` values depended on having a positive `productionTime` value. That is, when a `Robot` is placed on the map, it has some set time where it's doing something on the map.