Add `RouteFinder` class wrapping `MicroPather`
Use an explicit destructor in implementation file, to avoid implicit destructor instantiations in including files. The destructor will need to call the destructor for the `std::unique_ptr` field. Instantiating the `std::unique_ptr` destructor needs a complete type for the element being managed. Since the header file uses a forward declare, we want to ensure that instantiation happens in an explicit `RouteFinder` destructor in the corresponding implementation file.