Add `TileMapGraph` class to `RouteFinder.cpp`
The `TileMapGraph` class is a bit of an internal detail, though wasn't placed in an unnamed namespace. The header file has a forward declare for `TileMapGraph` so it can declare a `std::unique_ptr` field of that type. A type used in such a manner should not have internal linkage, and would produce a compile error if it did. Further, placing `TileMapGraph` in an unnamed namespace would change it's name, which needs to match the forward declare in the header file. As such, `TileMapGraph` is largely a private detail, though was not placed in an unnamed namespace with internal linkage.