Drop `const` restriction on `tile` in `SeedLander`
Ultimately the `tile` will be modified when the lander lands. Just right now it's done in a roundabout way by doing a lookup of the tile based on it's coordinates to get a `Tile&` with write access. We should remove the `const` so the roundabout lookup is no longer necessary.

By removing `const` we increase consistency with the other landers.