Remove error check that happens way too late
The `Tile` position should be validated when the robot is placed, either through user command, or from loading a saved game. There may already be dereferences of the `Tile` before this point, so if it wasn't validated, the game would have already crashed.

Additionally, the `getTile` call used for the offset position already has a built in check that would raise an exception for invalid `MapCoordinate`. That would be the case if the original `Tile` was too deep, unless someone an offset direction of `Up` was somehow used.