Update `CrimeRateUpdate::update()` to work without overlay
Rather than search an overlay list, search through a list of active police stations and range check against each. There will be a lot fewer entries in the active police stations collection than tiles in the overlay collection. It may actually be faster and take less memory to use the police station collection.

Additionally, the overlay list was more of a user interface object, which the core game engine shouldn't really know about. It's a bit odd to use it for both purposes.