Remove defaults from `Product` fields
The `Product` instances are always constructed with all fields initialized.

We should avoid setting default values for empty instances that are not usable. Better to not have to deal with zombie objects that don't have a clear purpose.

Removing the defaults reduces header uses of specific `ProductType` value names. That will be helpful for reducing header includes, though would be dependent on switching `ProductType` to `enum class` to enable forward declares.