Remove `ControlContainer::bringToFront`
This method was now only used internally, and was of questionable value. Controls are normally laid out so they don't overlap, which means `bringToFront` has no real use. The one part that may have been relevant was the `focus` change, though that tends to be handled by individual controls that can accept focus.