Add method `Tile::removeOreDeposit`
Make it an error to call `placeOreDeposit` on a `Tile` where `mOreDeposit` is already set. This isn't an issue for map generation, since the generation code ensures spacing between mines, so they'll never overlap. Potentially it could be an issue for a corrupt saved game file, though we should report such errors anyway.