Have `TileMap` own `OreDeposit` lifetimes
The `TileMap` creates the `OreDeposit` instances, and is the logical place to own and cleanup those instances.

This removes remaining uses of `delete` from `Tile`. The use of `delete` in `TileMap` is now paired with the `new` in `TileMap`.

This removes possible memory issues with the `Tile` copy members, which were explicitly deleted. It also removes the need to clear pointers in the `Tile` move members.