Convert `OreDeposit::serialize` to `static` method
The `serialize` method only needs to write fields with `public` accessor methods. That means it doesn't really need to be a member of `OreDeposit`. We can move the serialization code out of `OreDeposit`. That will be easier once this method is converted to `static` using only `public` methods, and so no longer makes assumptions about having direction access to private fields.