Return by-value from `structureList`
The internal detail of having structures broken up into lists by `StructureClass` will be removed. As such, there will be no internal list which can be returned by `const&`. For that reason we will build the list on demand. Additionally, the only external use of that method is called once per turn, during turn processing, so there isn't much of a performance impact to building the list on demand.