Update bio waste processing calculation
Allow waste processing to be concentrated on a first come first serve basis. We could have been smarter about this, such as prioritizing structures with overflow first, and perhaps ranking them so small amounts of overflow are processed first. That would mitigate most of the morale impacts. With that said, if we move to global storage for bio waste, there is little point in optimizing this now, when it will soon be replaced.