Simplify adding refined resources
Technically the comment may be slightly outdated, since storing in the Command Center first requires the structure list to be sorted such that Command Centers are first. This change relaxes that guarantee, since structures are now in their natural collection order, which is sorted once per turn to be in priority order. The priority order should place Command Centers first. However, there isn't a strong guarantee this method won't be called between modifying the structure list and sorting it in priority order. Thus there is a possibility a Command Center won't have priority for about the first turn it is built. Regardless, such a concern won't matter until we can build multiple Command Centers, and so there is a possibility of a Command Center being added later in the list. It also won't matter once we switch to global resources, which is likely to happen before we allow constructing multiple Command Centers.