Fix bad use of `std::clamp` in `ScrollBar::onThumbResize()`
We don't actually need `std::clamp` since the value can never exceed the given max. By using `std::max` we can avoid the undefined behavior in `std::clamp`. The max bound isn't needed since `naturalThumbLength` can never exceed the max bounds. When computing it, the max value is scaled according to `mLargeDelta` and `mMax`, so the scale expression is guaranteed to be `<= 1`, since `mLargeDelta <= (mMax + mLargeDelta)`, which we know because each of those values are clamped, with `mMax >= 0` and `mLargeDelta >= 1`.