Inject `OreHaulRoutes` into `MineReport`
This probably isn't a great design, though it does match existing practice for technology with the `ResearchReport`. The problem is that most core game objects are part of `MapViewState`, while `ReportsState` is not. We would need to address moving more data up to `GameState` instead of `MapViewState`, which is easier once `MapViewState` is split up, which this change helps with.