Update `ColonyShip` constructor to take available saved game data
This keeps use of the `ColonyShipOrbitTime` constant within the `ColonyShip` class itself, rather than in XML processing code. We should strive to keep such class specific constants contained within the class itself.

In general, classes should offer a constructor that allows initializing an object from data available in the saved game file. It doesn't necessarily need to be in the exact same format, though when conversion constants are needed, we should strive to keep them local to the class.